gPTP: Resolve warning: "comparison between signed and unsigned..."
Resolve compiler warning: "comparison between signed and unsigned
integer expressions." INVALID_LINKDELAY in avbts_port.hpp is defined
as a const int64_t, but is only ever compared against a uint64_t value.
Since it is const and initialized with a postive value, it should be
safe to make it an unsigned int.