Fixing a segfault that occurs when sending a PDelayReq.
Fixing a segfault that occurs when sending a PDelayReq while simultaneously
processing an unexpected PDelayRespFollowup. This can happen when running
multiple instances of gPTP over a non-802.1as-capable network switch that
forwards all layer two packets to every port. The thread that processes
incoming PDelayRespFollowup messages was deleting the port's pointer to the
last PDelayReq, which it was still in the process of sending. Repurposing an
existing lock to protect the last_pdelay_* pointers.