[gPTP] Prevent exit on EINTR in LinuxTimerQueue
Prevent LinuxTimerQueueHandler from exiting the timer thread when a
call to sigtimedwait returns EINTR. We have found that sigtimedwait
will occasionally return EINTR (indicating that the operation was
interrupted) while running tests that stress the CPU and flood the
network with UDP packets. Ignoring this error and retrying does not
appear to have ill effects. Change also includes:
   - Log unexpected errors from sigtimedwait (including EINTR).
   - Exit the program completely if the error from sigtimedwait is not
     EAGAIN (which indicates the timeout was reached) or EINTR. Leaving
     the program running in a non-functional state without the timer
     thread makes the failure less obvious than stopping completely.
   - Log errors related to the mutex in LinuxTimerQueueHandler and exit
     the program completely (same rationale as above).