Fixing deadlock when sending PDelayReq and receiving PDelayRespFup at the same time.
Fixing a deadlock that occurs when the PDelayReq timer expires while a valid
PDelayRespFup is being processed. The deadlock would occur because the two
threads would grab the same two locks (last_pdelay_lock and timerq_lock) but in
a different order.

The last_pdelay_lock is used to protect the last_pdelay_* variables which are
read/modified while sending a new PDelayReq and while processing a received
PDelayResp or PDelayRespFup to determine if the responses are valid.

The timerq_lock is held while adding/removing a timer from the queue.
E.g. #1: When the PDelayReq timer expires. It needs to be removed from the
queue and a new timer needs to be created after sending the new PDelayReq.
E.g. #2: When a valid PDelayRespFup has been processed so the PDelayRespFup
timer should be deleted.

In this case, the timerq_lock was being held for the entire duration of the
operation to send a new PDelayReq. Where as the thread processing an incoming
PDelayRespFup would only try to grab the timerq_lock if it needed to remove its
timer from the queue. So the two threads would grab the locks in reverse
order and deadlock. Fixing this by releasing the timerq_lock after removing the
expired timer from the queue but before performing the action associated with
it. This way, the timerq_lock is consistently used only to protect operations
on the timer queue and other locks within the processing functions (like
last_pdelay_lock) can be used to protect data or serialze operations in
whatever way makes sense for them.